Adding more unit tests and controller test

+++ src/main/java/com/presence/control/presenceapi/data/domain/Department.java
@@ -17,7 +17,7 @@ public class Department extends BaseEntity {
 
     private String departmentName;
 
-    private Long maxPeopleAllowed;
+    private Integer maxPeopleAllowed;
 
     @ManyToOne
     @JoinColumn(name = "department_id", referencedColumnName = "id")
